+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.1.1] - 2026-09-22
+
+### Added
+* Ability to query the package version update database
+
+### Improved
+* Prompt to update out-of-date Project Auditor package, combined with button to update the Rules package
+
 ## [3.1.0] - 2026-09-17
 
 ### Fixed
-* Improve the Obsolete attribute analyzer to ensure it detects everything from the obsolete database file.
-* Preserve UI selections across Domain Reloads.
+* Improve the Obsolete attribute analyzer to ensure it detects everything from the obsolete database file
+* Preserve UI selections across Domain Reloads
 
 ## [3.0.3] - 2026-09-02